requirements.txt
amqp==2.4.2
asn1crypto==0.24.0
beautifulsoup4==4.6.0
billiard==3.6.0.0
celery==4.3.0
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
cryptography==2.3
Django==1.11.20
django-audiofield==0.10.0
django-celery==3.2.2
django-filter==2.0.0
django-modelcluster==3.1
django-taggit==0.22.2
django-tagulous==0.14.0
django-tinymce==2.8.0
django-treebeard==4.3
djangorestframework==3.6.4
draftjs-exporter==2.1.5
enum34==1.1.6
gevent==1.2.2
greenlet==0.4.12
gunicorn==19.7.1
html5lib==0.999999999
idna==2.6
ipaddress==1.0.17
keyring==10.6.0
keyrings.alt==3.0
kombu==4.5.0
Markdown==3.1
netifaces==0.10.4
Pillow==5.4.1
psycopg2==2.7.4
psycopg2-binary==2.8
pycparser==2.19
pycrypto==2.6.1
pytz==2018.3
pyxdg==0.25
requests==2.21.0
SecretStorage==2.3.1
six==1.11.0
sqlparse==0.2.4
Unidecode==1.0.23
urllib3==1.24.1
vine==1.3.0
virtualenv==15.1.0
wagtail==1.13.4
wagtailmedia==0.2.0
webencodings==0.5.1
Willow==1.0
